Check visitor category when selecting banner
Don't show banners from campaigns where the campaign category matches a
category the visitor has already interacted with (closed, donated, etc).
Currently, the cookie is set by external sources (JavaScript).

Remove the concept of `hasDonated` from visitor and replace it with a
category list.
Add a test that checks if no banner is selected if the category matches.