Update PHP and Xdebug
Update PHP to 8 and Xdebug to 3

This update fixes our CI failures while building the Docker image. The
newest Debian version ("bullseye") has a problem when building PHP, so
we fix the underlying version to the previous one ("buster").

The CI environment seems to be fine, but we haven't tested the server
yet and it still runs on PHP 7.4 in production.

Our current Symfony version 4 might not be compatible with PHP 8 so this
is a preliminary commit to avoid the piling-up of failed dependabot pull
requests.