Update PHP and Xdebug
Update PHP to 8 and Xdebug to 3

This update is mostly to try if our CI failures while building the Docker
image go away. The CI environment seems to be fine, but we haven't
tested the server yet.

Our current Symfony version 4 might not be compatible with PHP 8 so this
is a preliminary commit to avoid the piling-up of failed dependabot pull
requests.